+++ Js2IL.Tests/Hosting/JavaScript/immutableExports.js
@@ -1,0 +1,13 @@
+"use strict";
+
+var exportsObject = {
+    lockedValue: 1
+};
+
+exportsObject.readLockedValue = function () {
+    return exportsObject.lockedValue;
+};
+
+Object.freeze(exportsObject);
+
+module.exports = exportsObject;